removed glamplify methods and cleaned up test to use Examples
